refactor(config-ui): move the offline page to error
