+++ config-ui/src/pages/index.ts
@@ -15,9 +15,7 @@
  * limitations under the License.
  *
  */
-
 export * from './error-handler'
-export * from './offline'
 export * from './project'
 export * from './blueprint'
 export * from './connections'
